+++ lib/diplomat/acl.rb
@@ -3,7 +3,7 @@
 module Diplomat
   # Methods for interacting with the Consul ACL API endpoint
   class Acl < Diplomat::RestClient
-    @access_methods = %i[list info create destroy update]
+    access_methods %i[list info create destroy update]
     attr_reader :id, :type, :acl
 
     # Get Acl info by ID