+++ lib/diplomat/event.rb
@@ -3,7 +3,7 @@
 module Diplomat
   # Methods for interacting with the Consul event API endpoint
   class Event < Diplomat::RestClient
-    @access_methods = %i[fire get_all get]
+    access_methods %i[fire get_all get]
 
     # Send an event
     # @param name [String] the event name